requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
drf-spectacular>=0.26.3
pillow
httpx==0.18.2
youtube-search-python
djangorestframework-simplejwt
channels
channels_redis
daphne